- issue19: Quit menu does not work - issue24: tree update is sometimes slow and makes all app unresponsive - issue23: add currentOp / killOp for each server - issue22: add nolock option to eval - issue7: randomize strings - issue4: compare collection sizes

+++ src/org/mongo/jmongob/CollectionNode.java
@@ -49,7 +49,7 @@ public class CollectionNode extends BaseTreeNode {
         label = collection.getName();
         CommandResult res = collection.getStats();
         res.throwOnError();
-        label += " (" + res.getInt("count") + "/" + res.getInt("size") + ")";
+        label += " (" + res.getLong("count") + "/" + res.getLong("size") + ")";
         if (res.getBoolean("sharded"))
             overlays.add(SwingFast.createIcon("overlay/superman.png", iconGroup));
     }
